I've started working on reimplementing the firmware from the disassembly.

+++ src/gpio.c
@@ -1,4 +1,5 @@
 #include "gpio.h"
+#include <stdint.h>
 
 /**
  * @brief Initialize the GPIO pins.
@@ -13,9 +14,9 @@ void gpio_init(uint32_t gpio_base, const gpio_config_t *config)
         if (config->pin & (1 << i)) {
             pin_pos = (i % 8) * 4;
             if (i < 8) {
-                cr = (uint32_t *)(gpio_base + 0x00);
+                cr = (uint32_t *)(uintptr_t)(gpio_base + 0x00);
             } else {
-                cr = (uint32_t *)(gpio_base + 0x04);
+                cr = (uint32_t *)(uintptr_t)(gpio_base + 0x04);
             }
             *cr &= ~((uint32_t)0xF << pin_pos);
             *cr |= ((uint32_t)(config->speed | config->mode) << pin_pos);
